requirements.txt
bcrypt==4.1.3
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
pymongo==4.7.1
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
